requirements.txt
aioboto3==14.1.0
aiobotocore==2.21.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aioitertools==0.12.0
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asyncpg==0.30.0
atlastk==0.13.4
attrs==25.3.0
backoff==2.2.1
boto3==1.37.1
botocore==1.37.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==45.0.2
Deprecated==1.2.18
distro==1.9.0
ecdsa==0.19.1
fastapi==0.115.11
frozenlist==1.5.0
googleapis-common-protos==1.69.2
grpcio==1.71.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib-metadata==6.0.1
iso8601==2.1.0
jiter==0.9.0
jmespath==1.0.1
jwt==1.3.1
multidict==6.2.0
openai==1.67.0
opentelemetry-api==1.17.0
opentelemetry-exporter-otlp==1.17.0
opentelemetry-exporter-otlp-proto-grpc==1.17.0
opentelemetry-exporter-otlp-proto-http==1.17.0
opentelemetry-instrumentation==0.38b0
opentelemetry-instrumentation-asgi==0.38b0
opentelemetry-instrumentation-fastapi==0.38b0
opentelemetry-instrumentation-httpx==0.38b0
opentelemetry-instrumentation-redis==0.38b0
opentelemetry-instrumentation-tortoiseorm==0.38b0
opentelemetry-proto==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-semantic-conventions==0.38b0
opentelemetry-util-http==0.38b0
packaging==25.0
passlib==1.7.4
propcache==0.3.1
protobuf==4.25.6
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyhumps==3.8.0
PyJWT==2.10.0
pyparsing==3.2.3
pypika-tortoise==0.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
redis==5.2.1
requests==2.32.3
rsa==4.9
s3transfer==0.11.3
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
tortoise==0.1.1
tortoise-orm==0.24.2
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
